Create layout for details fragment Move files
